2025-03-03 ZeroTier - health checking - alternative proposal
This PR follows on from the extensive discussion associated with #37.

Never before have I even *contemplated* submitting a PR covering the
same ground as an existing open PR. However, on this occasion I thought
it might be useful to have a concrete proposal to compare and contrast
with #37.

I sincerely hope that laying this on the (virtual) table and then
minimising further interaction *might* help us converge on a solution.

<hr>

Changes:

* `docker-compose.yml` and `docker-compose-router.yml`:

	- replaces deprecated `version` statement with `---`.

	- adds example environment variables.

* `Dockerfile`

	- corrects case of "as" to "AS" (silences build warning).

	- adds and configures `healthcheck.sh` (as per #37).

	- includes `tzdata` package (moved from `Dockerfile.router`) so
	  messages have local timestamps.

* `Dockerfile.router`

	- removes `tzdata` (moved to `Dockerfile`).

* `entrypoint-router.sh`:

	- code for first launch auto join of listed networks expanded to
	  include additional help material.

* `entrypoint.sh`:

	- "first launch" auto join of listed networks (code copied from
	  `entrypoint-router.sh`, as modified per above).

	- "self repair" of permissions in persistent store (code copied from
	  `entrypoint-router.sh`).

	- adds launch-time message to make it clear that the client is
	  launching (complements messages in `entrypoint-router.sh`).

	- abstracts some common strings to environment variables
	  (opportunistic change).

* `README.md`:

	- updates examples.

	- describes new environment variables (including move of
	  `ZEROTIER_ONE_NETWORK_IDS` from `README-router.md`.

	- documents health-checking.

* `README-router.md`

	- updates examples.
	- explains relationship of router and client.

Added:

* `healthcheck.sh`, based on original proposal in #37 and subsequent
  suggestions for modification by me.

I gave serious consideration to the code for synchronising networks in
the entry point scripts. The idea is quite attractive. It is safe to
automate joins in a "clean slate" situation. However, a *leave* followed
by a *join* is not guaranteed to be idempotent. That's because the
*leave* destroys the network-specific configuration options
(`allowManaged`, `allowGlobal`, `allowDefault`, `allowDNS`).

On balance I think it's better left to users to send explicit *leave*
commands via the CLI and take responsibility for restoring lost
configuration options on any subsequent *join*.

I will post the results of testing this PR separately.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>